Made it so that we can start making reporting work.
Currently, DBConnector has a SendQuery method that simply sends a query
and returns back a DataTable with allof teh data. We need to figure out
how to make the datatable display within WPF. I am hesitant to use a
datagrid, because it is so complex.

Right now, Patient lookup works. It was quite a bit of bodging, so I don't
expect you to understand how it works.